Renamed `com.contxt.streaming` package.
