+++ src/it/resources/reference.conf
@@ -1,9 +1,11 @@
-com.contxt.kinesis.consumer {
-  stats-class-name = "com.contxt.stream.NoopConsumerStats"
+com.contxt.kinesis {
+  consumer {
+    stats-class-name = "com.contxt.kinesis.NoopConsumerStats"
 
-  shard-checkpoint-config {
-    checkpoint-period = 60 seconds
-    checkpoint-after-completing-nr-of-records = 10000
-    max-wait-for-completion-on-stream-shutdown = 5 seconds
+    shard-checkpoint-config {
+      checkpoint-period = 60 seconds
+      checkpoint-after-completing-nr-of-records = 10000
+      max-wait-for-completion-on-stream-shutdown = 5 seconds
+    }
   }
 }
