+++ src/test/scala/com/contxt/kinesis/ShardCheckpointTrackerTest.scala
@@ -1,4 +1,4 @@
-package com.contxt.stream
+package com.contxt.kinesis
 
 import akka.Done
 import akka.util.ByteString
